@@ -5,22 +5,21 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.15.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_bloc: ^6.0.1
-  cached_network_image: ^2.3.3
-  equatable: ^1.2.4
-  dio: 3.0.9
-  shared_preferences: any
-  flutter_screenutil: ^4.0.3+1
-  flutter_flipperkit: ^0.0.24
-  flipperkit_sqflite_driver: ^0.0.2
-  flutter_svg: 0.19.1
-  fluttertoast: ^7.1.6
-  intl:
+  flutter_bloc: ^8.0.1
+  cached_network_image: ^3.2.0
+  equatable: ^2.0.3
+  dio: ^4.0.4
+  shared_preferences: ^2.0.12
+  flutter_screenutil: ^5.1.0
+  flutter_flipperkit: ^0.0.28
+  flutter_svg: ^1.0.3
+  fluttertoast: ^8.0.8
+  intl: ^0.17.0
 
 dev_dependencies:
   flutter_test:
