@@ -3,17 +3,17 @@ import 'package:warunkq_apps/helpers/app_color.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class BaseInput extends StatelessWidget {
-  final TextEditingController controller;
-  final String label;
-  final bool passwordField;
-  final TextInputType keyboardType;
-  final Widget suffixWidget;
-  final Function onClick;
-  final ValueChanged<String> validator;
-  final int maxLines;
-  final TextCapitalization textCapitalization;
-  final Color backgroundColor;
-  final ValueChanged<String> onChanged;
+  final TextEditingController? controller;
+  final String? label;
+  final bool? passwordField;
+  final TextInputType? keyboardType;
+  final Widget? suffixWidget;
+  final VoidCallback? onClick;
+  final FormFieldValidator<String>? validator;
+  final int? maxLines;
+  final TextCapitalization? textCapitalization;
+  final Color? backgroundColor;
+  final ValueChanged<String>? onChanged;
 
   BaseInput(
       {@required this.controller,
@@ -53,18 +53,18 @@ class BaseInput extends StatelessWidget {
       controller: controller,
       keyboardType: maxLines != null ? TextInputType.multiline : keyboardType,
       maxLines: maxLines,
-      obscureText: passwordField,
+      obscureText: passwordField!,
       textCapitalization: textCapitalization ?? TextCapitalization.none,
       readOnly: onClick != null,
       onChanged: onChanged,
-      validator: validator ?? (String args) => null,
-      style: TextStyle(fontSize: 16.sp, fontWeight: FontWeight.w500),
+      validator: validator,
+      style: TextStyle(fontSize: 14.sp, fontWeight: FontWeight.w500),
       decoration: InputDecoration(
           filled: backgroundColor != null ? true : false,
           hintText: label,
           hintStyle: TextStyle(
-              fontWeight: FontWeight.w500,
-              fontSize: 16.sp,
+              fontWeight: FontWeight.normal,
+              fontSize: 14.sp,
               color: AppColor.boxGrey),
           focusedBorder: OutlineInputBorder(
             borderRadius: BorderRadius.circular(15),
@@ -74,7 +74,7 @@ class BaseInput extends StatelessWidget {
             borderRadius: BorderRadius.circular(15),
             borderSide: BorderSide(color: AppColor.boxGrey, width: 1.0),
           ),
-          contentPadding: EdgeInsets.symmetric(vertical: 0.h, horizontal: 25.w),
+          contentPadding: EdgeInsets.symmetric(vertical: 0.h, horizontal: 15.w),
           fillColor: backgroundColor ?? Colors.transparent,
           focusColor: backgroundColor ?? Colors.transparent,
           suffixIcon: suffixWidget ?? Container(width: 2.w)),
