@@ -3,13 +3,13 @@ import 'package:warunkq_apps/helpers/app_color.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class StoreMenu extends StatelessWidget {
-  final String name;
-  final IconData icon;
-  final Function onTap;
-  final bool isLast;
+  final String? name;
+  final IconData? icon;
+  final VoidCallback? onTap;
+  final bool? isLast;
 
   const StoreMenu(
-      {Key key,
+      {Key? key,
       @required this.name,
       @required this.icon,
       @required this.onTap,
@@ -19,13 +19,13 @@ class StoreMenu extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
-      onTap: this.onTap,
+      onTap: onTap,
       child: Container(
         padding: EdgeInsets.symmetric(vertical: 15.h),
         decoration: BoxDecoration(
             color: Colors.white,
             border: Border(
-              bottom: !this.isLast ? BorderSide(color: AppColor.boxGrey.withOpacity(0.5), width: 1.5) : BorderSide(color: Colors.transparent),
+              bottom: !this.isLast! ? BorderSide(color: AppColor.boxGrey.withOpacity(0.5), width: 1.5) : BorderSide(color: Colors.transparent),
             )),
         child: Row(
           children: [
@@ -34,15 +34,15 @@ class StoreMenu extends StatelessWidget {
               child: Icon(
                 this.icon,
                 color: AppColor.black,
-                size: 20,
+                size: 18.sp,
               ),
             ),
             Expanded(
               child: Text(
-                this.name,
+                this.name!,
                 style: TextStyle(
                   color: AppColor.black,
-                  fontSize: 16.sp,
+                  fontSize: 14.sp,
                 ),
               ),
             ),
@@ -51,6 +51,7 @@ class StoreMenu extends StatelessWidget {
               child: Icon(
                 Icons.arrow_right_rounded,
                 color: AppColor.black,
+                size: 18.sp,
               ),
             ),
           ],
