@@ -2,19 +2,17 @@ import 'dart:io';
 
 import 'package:dio/dio.dart';
 import 'package:warunkq_apps/app.dart';
+import 'package:warunkq_apps/core/api.dart';
 import 'package:warunkq_apps/core/models/api_response.dart';
-import 'package:warunkq_apps/core/models/cart_cashier.dart';
 import 'package:warunkq_apps/core/models/transaction.dart';
 import 'package:warunkq_apps/helpers/url_helper.dart';
 
-class OrderAPI {
-  Future<ApiResponse<List<Transaction>>> findAll(
-    DateTime startDate,
-    DateTime endDate,
-  ) async {
+class OrderAPI implements API {
+  @override
+  Future<ApiResponse<List<Transaction>>> findAll(dates) async {
     try {
       Response res = await App().dio.get(UrlHelper.order,
-          queryParameters: {"startdate": startDate, "enddate": endDate});
+          queryParameters: {"startdate": dates.first, "enddate": dates.last});
       if (res.data['status'] == HttpStatus.ok) {
         return ApiResponse<List<Transaction>>(
             result: List.generate(
@@ -32,9 +30,10 @@ class OrderAPI {
     }
   }
 
-  Future<ApiResponse> create(CartCashier data) async {
+  @override
+  Future<ApiResponse<List<Transaction>>> create(data) async {
     try {
-      Response res = await App().dio.post(UrlHelper.order, data: data.toJson());
+      Response res = await App().dio.post(UrlHelper.order, data: data!.toJson());
       print(res.data);
       if (res.data['status'] == HttpStatus.ok) {
         return ApiResponse(
@@ -49,9 +48,10 @@ class OrderAPI {
     }
   }
 
-  Future<ApiResponse<List<Transaction>>> find(String orderNumber) async {
+  @override
+  Future<ApiResponse<List<Transaction>>> find(orderNumber) async {
     try {
-      Response res = await App().dio.get("${UrlHelper.order}/$orderNumber");
+      Response res = await App().dio.get("${UrlHelper.order}/${orderNumber.toString()}");
       print(res.data);
       if (res.data['status'] == HttpStatus.ok) {
         return ApiResponse<List<Transaction>>(
@@ -66,5 +66,17 @@ class OrderAPI {
       return ApiResponse(
           status: 500, message: "Something wrong, back in a minutes.");
     }
+  }
+
+  @override
+  Future<ApiResponse<List<Transaction>>> delete(dynamic) {
+    // TODO: implement delete
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<ApiResponse<List<Transaction>>> save(dynamic) {
+    // TODO: implement save
+    throw UnimplementedError();
   }
 }
